test: move related keyframes tests to feature spec
