+++ packages/core/test/stylable-transformer/scoping.spec.ts
@@ -1087,151 +1087,4 @@ describe('Stylable postcss transform (Scoping)', () => {
             testInlineExpects(result);
         });
     });
-
-    describe('@keyframes scoping', () => {
-        it('scope animation and animation name', () => {
-            const result = generateStylableRoot({
-                entry: `/entry.st.css`,
-                files: {
-                    '/entry.st.css': {
-                        namespace: 'entry',
-                        content: `
-                            /* @check entry__name */
-                            @keyframes name {
-                                from {}
-                                to {}
-                            }
-                            
-                            /* @check entry__name2 */
-                            @keyframes name2 {
-                                from {}
-                                to {}
-                            }
-
-                            /* @check .entry__selector {
-                                animation: 2s entry__name infinite, 1s entry__name2 infinite;
-                                animation-name: entry__name;
-                            }*/
-                            .selector {
-                                animation: 2s name infinite, 1s name2 infinite;
-                                animation-name: name;
-                            }
-
-                        `,
-                    },
-                },
-            });
-
-            testInlineExpects(result);
-        });
-
-        it('scope imported animation and animation name', () => {
-            const result = generateStylableRoot({
-                entry: `/entry.st.css`,
-                files: {
-                    '/entry.st.css': {
-                        namespace: 'entry',
-                        content: `
-                            :import {
-                                -st-from: './imported.st.css';
-                                -st-named: keyframes(anim1, anim2 as anim3);
-                            }
-                            /* @check .entry__selector {
-                                animation: 2s imported__anim1 infinite, 1s imported__anim2 infinite;
-                                animation-name: imported__anim1
-                            } */
-                            .selector {
-                                animation: 2s anim1 infinite, 1s anim3 infinite;
-                                animation-name: anim1;
-                            }
-
-                        `,
-                    },
-                    '/imported.st.css': {
-                        namespace: 'imported',
-                        content: `
-                            @keyframes anim1 {
-                                from {}
-                                to {}
-                            }
-
-                            @keyframes anim2 {
-                                from {}
-                                to {}
-                            }
-
-                        `,
-                    },
-                },
-            });
-
-            testInlineExpects(result);
-        });
-
-        it('scope imported animation and animation name with part name collision', () => {
-            const result = generateStylableRoot({
-                entry: `/entry.st.css`,
-                files: {
-                    '/entry.st.css': {
-                        namespace: 'entry',
-                        content: `
-                            :import {
-                                -st-from: './imported.st.css';
-                                -st-named: anim1, keyframes(anim1);
-                            }
-                            /* @check .entry__selector {
-                                animation: 2s imported__anim1 infinite;
-                                animation-name: imported__anim1;
-                            } */
-                            .selector {
-                                animation: 2s anim1 infinite;
-                                animation-name: anim1;
-                            }
-                            /* @check .imported__anim1 */
-                            .anim1{}
-                        `,
-                    },
-                    '/imported.st.css': {
-                        namespace: 'imported',
-                        content: `
-                            @keyframes anim1 {
-                                from {}
-                                to {}
-                            }
-                            .anim1 {}
-
-                        `,
-                    },
-                },
-            });
-            testInlineExpects(result);
-        });
-
-        it('not scope rules that are child of keyframe atRule', () => {
-            const result = generateStylableRoot({
-                entry: `/entry.st.css`,
-                files: {
-                    '/entry.st.css': {
-                        namespace: 'entry',
-                        content: `
-                            @keyframes name {
-                                /* @check from */
-                                from {}
-                                /* @check to */
-                                to {}
-                            }
-                            @keyframes name2 {
-                                /* @check 0% */
-                                0% {}
-                                /* @check 100% */
-                                100% {}
-                            }
-                        `,
-                    },
-                },
-            });
-
-            testInlineExpects(result);
-        });
-    });
 });
